Distribute pycmd also as a wheel
