+++ MANIFEST.in
@@ -1,0 +1,1 @@
+include src/pycmd/pycmd_public.html
