+++ README.txt
@@ -123,6 +123,7 @@ When reporting crashes, please try to locate and attach a crash log (look in
             http://code.activestate.com/recipes/146262
 
 
-
 ---------------------------------------------------
 Horea Haitonic (h o r e a h _at_ g m a i l . c o m)
+---------------------------------------------------
+
