+++ pyproject.toml
@@ -1,0 +1,79 @@
+[build-system]
+requires = ["setuptools"]
+build-backend = "setuptools.build_meta"
+
+[project]
+authors = [
+  {name = "Horea Haitonic", email = "horeah@gmail.com"},
+]
+classifiers = [
+  "Development Status :: 5 - Production/Stable",
+  "Environment :: Console",
+  "Intended Audience :: Developers",
+  "Intended Audience :: System Administrators",
+  "Operating System :: Microsoft :: Windows",
+  "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
+  "Programming Language :: Python :: 3.13",
+  "Programming Language :: Python :: 3.14",
+  "Topic :: System :: Shells",
+  "Topic :: Terminals",
+  "Topic :: Utilities",
+]
+dependencies = [
+  "patchelf; platform_system!='Windows'",
+  "pefile; platform_system=='Windows'",
+  "pywin32; platform_system=='Windows'",
+  "chatlas",
+  "google",
+]
+description = "Improved interactive shell"
+dynamic = ["version"]
+keywords = ["cmd"]
+# "license" property workaround to make both setuptools and cx_freeze happy:
+# 1. setuptools actually already requires PEP 639 (license = <str> or
+#    license-files = []), so it prints deprecation warnings. The LICENSE.txt
+#    gets auto-included anyway.
+# 2. cx_freeze still expects an older PEP, requiring license = {text = <str>}.
+#    We include LICENSE.txt via 'include_files'.
+license = {text = ""}
+maintainers = [
+  {name = "Horea Haitonic", email = "horeah@gmail.com"},
+]
+name = "py-cmd"
+readme = "README-WHL.txt"
+requires-python = ">=3.10"
+
+[project.scripts]
+pycmd = "pycmd.PyCmd:_main"
+
+[project.urls]
+Changelog = "https://github.com/horeah/PyCmd/blob/master/NEWS.txt"
+Documentation = "https://github.com/horeah/PyCmd/blob/master/README.txt"
+Homepage = "https://github.com/horeah/PyCmd"
+Issues = "https://github.com/horeah/PyCmd/issues"
+Repository = "https://github.com/horeah/PyCmd"
+
+[tool.setuptools.dynamic]
+version = {attr = "pycmd.buildinfo.build_date"}
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+[tool.cxfreeze]
+executables = [
+    {script = "PyCmd.py", icon = "PyCmd.ico"}
+]
+
+[tool.cxfreeze.build_exe]
+excludes = ["lib2to3", "Tkinter", "Tk", "Tcl", "test"]
+include_files = [
+    "LICENSE.txt",
+    "README.txt",
+    "src/pycmd/PyCmd.py",
+    "src/pycmd/example-init.py",
+    "src/pycmd/pycmd_public.html",
+]
+packages = ["chatlas", "googlesearch"]
