+++ src/pycmd/DirHistory.py
@@ -1,9 +1,9 @@
 import os, sys
 import os.path
-from console import get_cursor, move_cursor, get_buffer_size
+from pycmd.console import get_cursor, move_cursor, get_buffer_size
 from sys import stdout
-from common import abbrev_tilde
-from pycmd_public import appearance, color
+from pycmd.common import abbrev_tilde
+from pycmd.pycmd_public import appearance, color
 
 def _norm(path):
     return os.path.normcase(os.path.expanduser(path))
