+++ src/pycmd/__main__.py
@@ -1,0 +1,3 @@
+from .PyCmd import _main
+
+_main()
