+++ src/pycmd/completion.py
@@ -5,9 +5,9 @@
 #
 
 import sys, os, re, time
-from common import tokenize, expand_env_vars, has_exec_extension, is_executable, strip_extension
-from common import contains_special_char, starts_with_special_char
-from common import sep_chars, seq_tokens
+from pycmd.common import tokenize, expand_env_vars, has_exec_extension, is_executable, strip_extension
+from pycmd.common import contains_special_char, starts_with_special_char
+from pycmd.common import sep_chars, seq_tokens
 
 def complete_universal(line):
     """
