+++ src/pycmd/console/console_common.py
