+++ src/pycmd/fsm.py
